change EncryptAlgorithm in PasswordStoreKeyInsertDto ulong to int

+++ cloudsharpback/Models/PasswordStoreKeyInsertDto.cs
@@ -2,7 +2,7 @@ namespace cloudsharpback.Models;
 
 public record PasswordStoreKeyInsertDto
 {
-    public required ulong EncryptAlgorithm { get; init; }   
+    public required int EncryptAlgorithm { get; init; }   
     public string? PublicKey { get; init; }   
     public required string PrivateKey { get; init; }   
 }
