+++ cloudsharpback/Repository/PasswordStoreKeyRepository.cs
@@ -29,7 +29,7 @@ WHERE password_store_key_id = @keyId;
         return await conn.QueryFirstOrDefaultAsync(sql, new { keyId });
     }
 
-    public async Task<bool> InsertKey(ulong memberId, ulong encryptAlgorithm, string? publicKey, string privateKey)
+    public async Task<bool> InsertKey(ulong memberId, int encryptAlgorithm, string? publicKey, string privateKey)
     {
         const string sql = @"
 Insert Into password_store_keys(owner_id, encrypt_algorithm, public_key, private_key)
